fix: develop 브랜치와 feature-56 병합 충돌 해결

+++ BE/src/main/java/com/codaily/project/service/TaskDetailServiceImpl.java
@@ -226,7 +226,7 @@ public class TaskDetailServiceImpl implements TaskDetailService {
                 .description(feature.getDescription())
                 .category(feature.getCategory())
                 .priorityLevel(feature.getPriorityLevel())
-                .estimatedTime(feature.getEstimatedTime())
+                .estimatedTime(feature.getEstimatedTime().intValue())
                 .build();
     }
